[#2] split module
